Update package silex-wlanbt-fs to v4.5.25.38-0
The new driver version is based on version 4.5.25.38 of the original
Qualcomm driver which is available in branch CNSS.LEA.NRT_3.1 on
repository (tag v4.5.25.38)

  http://source.codeaurora.org/external/wlan/qcacld-2.0

But Silex has added a whole lot of commits that were done later in
the development, up to version 4.5.25.52. And Silex has also added
some own changes.

Through a detailed analysis of the changes we were able to define
a complete list of patches that Silex has taken from the original
code. Then we added one final patch that brings in all Silex specific
changes. This results in the 56 patches that are now here in the
F&S package. Of course all previous patches are obsolete now.

So download the right version from the codeaurora git repository,
and apply alle the patches. This results in the Silex version. The
full official Silex version is "4.5.25.38 sxa 1.0.0.019".