Move S02silex from silex-wlanbt-fs to silex-wlanbt-fs-firmware
The Silex WLAN module needs a small hack to allow a dynamic MAC address
that is set in U-Boot. The startup script S02silex does the necessary setup.
However it references directory /lib/firmware/wlan. This directory is
created by package silex-wlanbt-fs-firmware. But the script was part
of package silex-wlanbt-fs, which is a bit illogical. So move the
installation of this script to silex-wlanbt-fs-firmware, where it
makes more sense.