Update package silex-wlanbt-fs to 4.5.25.38-1
The new driver version is based on version 4.5.25.38 of the original
Qualcomm driver which is available in branch CNSS.LEA.NRT_3.1 on
repository (tag v4.5.25.38)

  https://git.codelinaro.org/clo/la/platform/vendor/qcom-opensource/wlan/qcacld-2.0

Also added is a set of commits done later in the development, up
to version 4.5.25.56. Through a detailed analysis of the changes
we were able to define a complete list of patches that Silex has
taken from the original code.

This results a Silex-free version of the driver, just consisting
of code and those 18 patches from the above qcacld-2.0 tree,
including a patch to handle newer kernel versions up to 6.1.
Please note that the set of patches differs slightly to those
of 4.5.25.38-0, which means all previous patches are obsolete now.

There is also a Silex-specific version availabe on request that
improves roaming, adds bang radar and other improvements. Ask
F&S if interested.