Update silex-wlanbt-fs-firmware to v4.5.25.38-1
The new Silex driver also needs a new set of firmware files.

/lib/firmware/WLAN-firmware/bin/bdwlan30.bin:
   F&S sepcific board config, same as before
/lib/firmware/WLAN-firmware/bin/qwlan30.bin:
   New firmware from the Qualcomm driver (firmware/WLAN-firmware/bin)
/lib/firmware/WLAN-firmware/bin/otp30.bin:
   Same as before (drivers/firmware/WLAN-firmware/otp30.bin)
/lib/firmware/utf30.bin:
   Same as before (drivers/firmware/WLAN-firmware/bin)
/lib/firmware/utfbd30.bin:
   Same as bdwlan30.bin (symbolic link)
/lib/firmware/wlan/qcom_cfg.ini:
   New Silex ini file, from original Qualcomm version, only
   minor changes (qcacld-new/firmware_bin/qcom_cfg.ini)
/lib/firmware/wlan/cfg.dat:
   Same as before (renamed from qcacld-new/firmware_bin/WCNSS_cfg.dat)

/lib/firmware/BT-firmware/BD-SDMAC/nvm_tlv_tf_1.1.bin:
   F&S specific version, no changes
/lib/firmware/qca/rampatch_le_2-3-3_tlv_tf_1.1_bt4-2.tlv
   Like before, support for Bluetooth-4.2, no changes
/lib/firmware/qca/rampatch_tlv_tf_1.1.tlv:
   Same as above (symbolic link)

There is also a new regulatory database from Silex. The driver maps
radio signal strengths of the different countries/domains taken from
the regulatory database to a table located in the board specific file
bdwlan30.bin. Unfortunately, the table has only a limited number of
entries. This is why Silex combines rather similar countries in their
database to use the same settings. For example if countries X and Y
only differ on one channel by 1dB, then they they are collapsed to
one entry, using the more restrictive setting of the two.
This saves a table entry. The resulting settings can be used in both
countries, it just transmits a little bit weaker than necessary in
one of the two countries on this one channel.

/lib/firmware/regulatory.db:
   Silex specifc database, from Silex patch 2024-10-30 (regulatory.db)
   (an older version is in drivers/firmware/regulatory.db)
/lib/firmware/regulatory.db.p7s:
   Silex regdb signature (drivers/firmware/regulatory.db.p7s)

The firmware files themselves are provided in the new tar file

  silex-wlanbt-fs-firmware-v4.5.25.38-1.tar.bz2

that has to be added to all releases where the Silex chip is part of. So
this commit only changes the makefile silex-wlanbt-fs-firmware.mk.